% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conv.mean.R
\name{conv.mean}
\alias{conv.mean}
\title{Running mean function}
\usage{
conv.mean(x)
}
\arguments{
\item{x}{a variable whose convergence we want to check}
}
\value{
returns a running mean of the variable whose convergence we want to check
}
\description{
A running mean, intended to check when convergence is achieved (useful for setting mcsize or bssize).
}
\examples{
\dontrun{
place example here!!
}
}
